refactor: move out spi project
