+++ service-framework-spi/build.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+  `java-library`
+  id("org.hypertrace.publish-plugin")
+}
